gambit2/gambit2/requirements.txt
paramiko
docker
pydantic
asyncio
python-dotenv
google-genai
